+++ e2e/config-noesm-js/cypress/support/e2e.js
@@ -1,0 +1,9 @@
+const registerCypressGrep = require("@cypress/grep/src/support");
+require("cypress-terminal-report/src/installLogsCollector")();
+require("cypress-cloud/support");
+require("./commands");
+
+registerCypressGrep();
+beforeEach(() => {
+  cy.visit("/");
+});