+++ examples/webapp/currents.config.js
@@ -1,12 +1,13 @@
 module.exports = {
   e2e: {
-    batchSize: 3, // how many specs to send in one batch
+    batchSize: 2, // how many specs to send in one batch
   },
   component: {
     batchSize: 5, // how many specs to send in one batch
   },
   projectId: !!(process.env.GITHUB_ACTION || process.env.CIRCLE_BRANCH)
     ? "Ij0RfK"
-    : "l4zuz8",
+    : "Ij0RfK",
   // cloudServiceUrl: "http://localhost:1234",
+  userAgent: "custom",
 };