+++ packages/cypress-cloud/lib/config/config.ts
@@ -1,12 +1,12 @@
 import crypto from 'crypto';
 
 import Debug from "debug";
-import path from "path";
+
+import { P, match } from "ts-pattern";
 import { DetectedBrowser, ValidatedCurrentsParameters } from "../../types";
 import { bootCypress } from "../bootstrap";
-import { warn } from "../log";
-import { require } from "../require";
-import { getRandomPort } from "../utils";
+import { info, warn } from "../log";
+import { getConfigFilePath } from "./path";
 
 const debug = Debug("currents:config");
 
@@ -20,13 +20,28 @@ export type CurrentsConfig = {
   cloudServiceUrl: string;
   component: ComponentConfig;
   e2e: E2EConfig;
+  networkHeaders?: Record<string, string>;
   projectId?: string;
   recordKey?: string;
 };
 
 let _config: CurrentsConfig | null = null;
 
-export function getCurrentsConfig(projectRoot?: string): CurrentsConfig {
+const defaultConfig: CurrentsConfig = {
+  e2e: {
+    batchSize: 3,
+  },
+  component: {
+    batchSize: 5,
+  },
+  cloudServiceUrl: "https://cy.currents.dev",
+  networkHeaders: undefined,
+};
+
+export async function getCurrentsConfig(
+  projectRoot?: string,
+  explicitConfigFilePath?: string
+): Promise<CurrentsConfig> {
   if (_config) {
     return _config;
   }
@@ -40,19 +55,39 @@ export function getCurrentsConfig(projectRoot?: string): CurrentsConfig {
     cloudServiceUrl: "https://api.deploysentinel.com",
   };
 
-  const configFilePath = getConfigFilePath(projectRoot);
-  try {
-    const resolvedPath = path.resolve(...configFilePath);
-    debug("loading currents config file from '%s'", resolvedPath);
+  const configFilePath = getConfigFilePath(projectRoot, explicitConfigFilePath);
+  // try loading possible config files
+  for (const filepath of configFilePath) {
+    const config = match(await loadConfigFile(filepath))
+      .with({ default: P.not(P.nullish) }, (c) => c.default)
+      .with(P.not(P.nullish), (c) => c)
+      .otherwise(() => null);
 
-    const fsConfig = require(resolvedPath);
-    _config = {
-      ...defaultConfig,
-      ...fsConfig,
-    } as CurrentsConfig;
-    return _config;
+    if (config) {
+      debug("loaded currents config from '%s'\n%O", filepath, config);
+      info("Using config file: '%s'", filepath);
+      _config = {
+        ...defaultConfig,
+        ...config,
+      };
+      return _config;
+    }
+  }
+
+  warn(
+    "Failed to load config file, falling back to the default config. Attempted locations: %s",
+    configFilePath
+  );
+  _config = defaultConfig;
+  return _config;
+}
+
+async function loadConfigFile(filepath: string) {
+  try {
+    debug("loading currents config file from '%s'", filepath);
+    return await import(filepath);
   } catch (e) {
-    debug("failed to load config file: %s", configFilePath);
+    warn("failed to load config file: %s", filepath);
     debug("failure details: %s", e);
     _config = defaultConfig;
     return _config;
@@ -68,7 +103,7 @@ export async function getMergedConfig(params: ValidatedCurrentsParameters) {
         rawJson: Record<string, unknown>;
         browsers: DetectedBrowser[];
       })
-    | undefined = await bootCypress(getRandomPort(), params);
+    | undefined = await bootCypress(params);
 
   debug("cypress resolvedConfig: %O", cypressResolvedConfig);
 
@@ -94,8 +129,4 @@ export async function getMergedConfig(params: ValidatedCurrentsParameters) {
   };
   debug("merged config: %O", result);
   return result;
-}
-
-function getConfigFilePath(projectRoot: string | null = null) {
-  return [projectRoot ?? process.cwd(), "currents.config.js"];
 }