+++ packages/cypress-cloud/lib/run.ts
@@ -4,7 +4,7 @@ import crypto from "crypto";
 import Debug from "debug";
 import { CurrentsRunParameters } from "../types";
 import { createRun } from "./api";
-import { cutInitialOutput } from "./capture";
+import { cutInitialOutput, getCapturedOutput } from "./capture";
 import { getCI } from "./ciProvider";
 import {
   getMergedConfig,
@@ -13,13 +13,26 @@ import {
   validateParams,
 } from "./config";
 import { runBareCypress } from "./cypress";
+import { activateDebug } from "./debug";
 import { getGitInfo } from "./git";
 import { setAPIBaseUrl, setRunId } from "./httpClient";
 import { bold, divider, info, spacer, title } from "./log";
 import { getPlatform } from "./platform";
+import { pubsub } from "./pubsub";
 import { summarizeTestResults, summaryTable } from "./results";
-import { runTillDoneOrCancelled, summary, uploadTasks } from "./runner";
+import {
+  createReportTaskSpec,
+  getExecutionStateResults,
+  reportTasks,
+  runTillDoneOrCancelled,
+  setConfig,
+  setSpecAfter,
+  setSpecBefore,
+  setSpecOutput,
+} from "./runner";
+import { shutdown } from "./shutdown";
 import { getSpecFiles } from "./specMatcher";
+import { startWSS } from "./ws";
 
 const debug = Debug("currents:run");
 
@@ -52,6 +65,7 @@ const setDSCustomEnvs = () => {
 };
 
 export async function run(params: CurrentsRunParameters = {}) {
+  activateDebug(params.cloudDebug);
   debug("run params %o", params);
   params = preprocessParams(params);
   debug("params after preprocess %o", params);
@@ -60,7 +74,8 @@ export async function run(params: CurrentsRunParameters = {}) {
     info(`Skipping cloud orchestration because --record is set to false`);
     return runBareCypress(params);
   }
-  const validatedParams = validateParams(params);
+
+  const validatedParams = await validateParams(params);
   setAPIBaseUrl(validatedParams.cloudServiceUrl);
 
   setDSCustomEnvs();
@@ -78,6 +93,10 @@ export async function run(params: CurrentsRunParameters = {}) {
   } = validatedParams;
 
   const config = await getMergedConfig(validatedParams);
+
+  // %state
+  setConfig(config?.resolved);
+
   const { specs, specPattern } = await getSpecFiles({
     config,
     params: validatedParams,
@@ -91,8 +110,6 @@ export async function run(params: CurrentsRunParameters = {}) {
     config,
     browser: validatedParams.browser,
   });
-
-  divider();
 
   info("Discovered %d spec files", specs.length);
   info(
@@ -119,13 +136,16 @@ export async function run(params: CurrentsRunParameters = {}) {
     autoCancelAfterFailures,
   });
 
-  info("🎥 Run URL:", bold(run.runUrl));
-
   setRunId(run.runId);
 
   setDSMachineId(run.machineId);
 
+  info("🎥 Run URL:", bold(run.runUrl));
+
   cutInitialOutput();
+
+  await startWSS();
+  listenToSpecEvents();
 
   await runTillDoneOrCancelled(
     {
@@ -133,7 +153,6 @@ export async function run(params: CurrentsRunParameters = {}) {
       groupId: run.groupId,
       machineId: run.machineId,
       platform,
-      config,
       specs,
     },
     validatedParams
@@ -141,12 +160,14 @@ export async function run(params: CurrentsRunParameters = {}) {
 
   divider();
 
-  await Promise.allSettled(uploadTasks);
-  const _summary = summarizeTestResults(Object.values(summary), config);
+  await Promise.allSettled(reportTasks);
+  const _summary = summarizeTestResults(getExecutionStateResults(), config);
 
   title("white", "Cloud Run Finished");
   console.log(summaryTable(_summary));
   info("🏁 Recorded Run:", bold(run.runUrl));
+
+  await shutdown();
 
   spacer();
   if (_summary.status === "finished") {
@@ -155,5 +176,23 @@ export async function run(params: CurrentsRunParameters = {}) {
       runUrl: run.runUrl,
     };
   }
+
   return _summary;
+}
+
+function listenToSpecEvents() {
+  pubsub.on("before:spec", async ({ spec }: { spec: Cypress.Spec }) => {
+    debug("before:spec %o", spec);
+    setSpecBefore(spec.relative);
+  });
+
+  pubsub.on(
+    "after:spec",
+    async ({ spec, results }: { spec: Cypress.Spec; results: any }) => {
+      debug("after:spec %o %o", spec, results);
+      setSpecAfter(spec.relative, results);
+      setSpecOutput(spec.relative, getCapturedOutput());
+      createReportTaskSpec(spec.relative);
+    }
+  );
 }