+++ packages/cypress-cloud/lib/runner/runner.ts
@@ -3,13 +3,8 @@ import {
   ValidatedCurrentsParameters,
 } from "@deploysentinel/cypress-parallel/types";
 import { getCapturedOutput, resetCapture } from "../capture";
-import { MergedConfig } from "../config";
 
-import {
-  getSummaryForSpec,
-  getUploadResultsTask,
-  normalizeRawResult,
-} from "../results";
+import { getCypressRunResultForSpec } from "../results";
 
 import Debug from "debug";
 import {
@@ -21,8 +16,13 @@ import {
 
 import { runSpecFileSafe } from "../cypress";
 import { isCurrents } from "../env";
-import { divider, error, info, title, warn } from "../log";
-import { summary, uploadTasks } from "./state";
+import { divider, info, title, warn } from "../log";
+import { createReportTask, reportTasks } from "./reportTask";
+import {
+  initExecutionState,
+  setInstanceOutput,
+  setInstanceResult,
+} from "./state";
 
 const debug = Debug("currents:runner");
 
@@ -32,10 +32,8 @@ export async function runTillDone(
     groupId,
     machineId,
     platform,
-    config,
     specs: allSpecs,
   }: CreateInstancePayload & {
-    config: MergedConfig;
     specs: SpecWithRelativeRoot[];
   },
   params: ValidatedCurrentsParameters
@@ -52,25 +50,18 @@ export async function runTillDone(
       },
       allSpecs,
       params,
-      config,
     });
     if (!newTasks.length) {
-      debug("No more tasks to run. Uploads queue: %d", uploadTasks.length);
+      debug("No more tasks to run. Uploads queue: %d", reportTasks.length);
       hasMore = false;
       break;
     }
-    newTasks.forEach((task) => {
-      if (task.summary.specSummary) {
-        summary[task.summary.spec] = task.summary.specSummary;
-      }
-      uploadTasks.push(task.uploadTasks);
-    });
+    newTasks.forEach((t) => createReportTask(t.instanceId));
   }
 }
 
 async function runBatch({
   runMeta,
-  config,
   params,
   allSpecs,
 }: {
@@ -81,7 +72,6 @@ async function runBatch({
     platform: CreateInstancePayload["platform"];
   };
   allSpecs: SpecWithRelativeRoot[];
-  config: MergedConfig;
   params: ValidatedCurrentsParameters;
 }) {
   let batch = {
@@ -114,6 +104,25 @@ async function runBatch({
     return [];
   }
 
+  /**
+   * Batch can have multiple specs. While running the specs,
+   * cypress can hard-crash without reporting any result.
+   *
+   * When crashed, ideally, we need to:
+   * - determine which spec crashed
+   * - associate the crash with the spec
+   * - run the rest of unreported specs in the batch
+   *
+   * But detecting the crashed spec is error-prone and inaccurate,
+   * so we fall back to reporting hard crash to all subsequent
+   * specs in the batch.
+   *
+   * Worst-case scenario: we report hard crash to all specs in the batch.
+   */
+
+  // %state
+  batch.specs.forEach(initExecutionState);
+
   divider();
   info(
     "Running: %s (%d/%d)",
@@ -124,45 +133,32 @@ async function runBatch({
 
   const rawResult = await runSpecFileSafe(
     {
-      // use absolute paths -  user can run the program from a different directory, e.g. nx or a monorepo workspace
-      // cypress still reports the path relative to the project root
+      // use absolute paths - user can run the program from a different directory, e.g. nx or a monorepo workspace
+      // cypress still report the path relative to the project root
       spec: batch.specs
         .map((bs) => getSpecAbsolutePath(allSpecs, bs.spec))
         .join(","),
     },
     params
   );
-  const normalizedResult = normalizeRawResult(
-    rawResult,
-    batch.specs.map((s) => s.spec),
-    config
-  );
 
   title("blue", "Reporting results and artifacts in background...");
 
   const output = getCapturedOutput();
-  resetCapture();
 
-  const batchResult = batch.specs.map((spec) => {
-    const specSummary = getSummaryForSpec(spec.spec, normalizedResult);
-    if (!specSummary) {
-      warn('Cannot find run result for spec "%s"', spec.spec);
+  // %state
+  batch.specs.forEach((spec) => {
+    setInstanceOutput(spec.instanceId, output);
+    const specRunResult = getCypressRunResultForSpec(spec.spec, rawResult);
+    if (!specRunResult) {
+      return;
     }
-
-    return {
-      summary: {
-        spec: spec.spec,
-        specSummary,
-      },
-      uploadTasks: getUploadResultsTask({
-        ...spec,
-        runResult: normalizedResult,
-        output,
-      }).catch(error),
-    };
+    setInstanceResult(spec.instanceId, specRunResult);
   });
 
-  return batchResult;
+  resetCapture();
+
+  return batch.specs;
 }
 
 function getSpecAbsolutePath(