+++ packages/cypress-cloud/lib/shutdown.ts
@@ -1,0 +1,5 @@
+import { stopWSS } from "./ws";
+
+export async function shutdown() {
+  await stopWSS();
+}