+++ packages/cypress-cloud/lib/ws.ts
@@ -1,2 +1,0 @@
-// plug for future websocket server
-export {};